--- compatibility-api/src/main/scala/za/co/absa/hyperdrive/compatibility/api/CompatibleDeltaCDCToSCD2Writer.scala
@@ -1,37 +1,0 @@
-/*
- * Copyright 2018 ABSA Group Limited
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package za.co.absa.hyperdrive.compatibility.api
-
-import org.apache.spark.sql.DataFrame
-import org.apache.spark.sql.streaming.{StreamingQuery, Trigger}
-
-trait CompatibleDeltaCDCToSCD2Writer {
-  def write(dataFrame: DataFrame): StreamingQuery
-}
-
-case class DeltaCDCToSCD2WriterConfiguration(
-  destination: String,
-  trigger: Trigger,
-  checkpointLocation: String,
-  partitionColumns: Seq[String],
-  keyColumn: String,
-  timestampColumn: String,
-  operationColumn: String,
-  operationDeleteValues: Seq[String],
-  precombineColumns: Seq[String],
-  precombineColumnsCustomOrder: Map[String, Seq[String]],
-  extraConfOptions: Map[String, String]
-)
