+++ compatibility_spark-3/src/main/scala/za/co/absa/hyperdrive/compatibility/impl/writer/delta/scd2/DeltaCDCToSCD2WriterAttributes.scala
@@ -13,7 +13,7 @@
  * limitations under the License.
  */
 
-package za.co.absa.hyperdrive.ingestor.implementation.writer.deltacdctoscd2writer
+package za.co.absa.hyperdrive.compatibility.impl.writer.delta.scd2
 
 import za.co.absa.hyperdrive.ingestor.api.writer.StreamWriterCommonAttributes
 import za.co.absa.hyperdrive.ingestor.api.{HasComponentAttributes, PropertyMetadata}
