+++ compatibility_spark-3/src/test/scala/za/co/absa/hyperdrive/compatibility/impl/writer/delta/scd2/DeltaCDCToSCD2Classes.scala
@@ -13,7 +13,7 @@
  * limitations under the License.
  */
 
-package za.co.absa.hyperdrive.compatibility.impl.TestDeltaCDCToSCD2Writer
+package za.co.absa.hyperdrive.compatibility.impl.writer.delta.scd2
 
 import za.co.absa.hyperdrive.compatibility.impl.writer.delta.FileUtils
