--- ingestor-default/src/main/scala/za/co/absa/hyperdrive/ingestor/implementation/writer/deltacdctoscd2writer/DeltaCDCToSCD2WriterLoader.scala
@@ -1,22 +1,0 @@
-/*
- * Copyright 2018 ABSA Group Limited
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package za.co.absa.hyperdrive.ingestor.implementation.writer.deltacdctoscd2writer
-
-import za.co.absa.hyperdrive.ingestor.api.writer.{StreamWriterFactory, StreamWriterFactoryProvider}
-
-class DeltaCDCToSCD2WriterLoader extends StreamWriterFactoryProvider {
-  override def getComponentFactory: StreamWriterFactory = DeltaCDCToSCD2Writer
-}
